refactor: move ICipherPayloadHandler interface to a separate file for better organization
