+++ Ecliptix.Core/Protocol/CipherPayloadHandler.cs
@@ -8,22 +8,6 @@ using Grpc.Core;
 
 namespace Ecliptix.Core.Protocol;
 
-public interface ICipherPayloadHandler
-{
-    Task<Result<CipherPayload, FailureBase>> EncryptResponse(byte[] payload, uint connectId, ServerCallContext context);
-    Task<Result<byte[], FailureBase>> DecryptRequest(CipherPayload cipherPayload, uint connectId, ServerCallContext context);
-
-    Task<CipherPayload> RespondSuccess<T>(byte[] payload, uint connectId, ServerCallContext context)
-        where T : IMessage<T>, new();
-
-    Task<CipherPayload> RespondFailure<T>(FailureBase failure, uint connectId, ServerCallContext context)
-        where T : IMessage<T>, new();
-    
-    Task<CipherPayload> HandleResult<TSuccess, TFailure>(Result<TSuccess, TFailure> result, uint connectId, ServerCallContext context)
-        where TSuccess : IMessage<TSuccess>, new()
-        where TFailure : FailureBase;
-}
-
 public class CipherPayloadHandler : ICipherPayloadHandler
 {
     private readonly IActorRef _protocolActor;
